deler opp motta i behandling og vedtak
