+++ src/main/kotlin/no/nav/tiltakspenger/datadeling/identhendelse/IdenthendelseService.kt
@@ -1,8 +1,8 @@
 package no.nav.tiltakspenger.datadeling.identhendelse
 
 import io.github.oshai.kotlinlogging.KotlinLogging
-import no.nav.tiltakspenger.datadeling.motta.infra.db.BehandlingRepo
-import no.nav.tiltakspenger.datadeling.motta.infra.db.VedtakRepo
+import no.nav.tiltakspenger.datadeling.motta.behandling.db.BehandlingRepo
+import no.nav.tiltakspenger.datadeling.motta.vedtak.db.VedtakRepo
 import no.nav.tiltakspenger.libs.common.Fnr
 import java.util.UUID
