+++ src/main/kotlin/no/nav/tiltakspenger/datadeling/service/BehandlingService.kt
@@ -3,7 +3,7 @@ package no.nav.tiltakspenger.datadeling.service
 import no.nav.tiltakspenger.datadeling.domene.Behandling
 import no.nav.tiltakspenger.datadeling.domene.Kilde
 import no.nav.tiltakspenger.datadeling.domene.TiltakspengerBehandling
-import no.nav.tiltakspenger.datadeling.motta.infra.db.BehandlingRepo
+import no.nav.tiltakspenger.datadeling.motta.behandling.db.BehandlingRepo
 import no.nav.tiltakspenger.libs.common.Fnr
 import no.nav.tiltakspenger.libs.periodisering.Periode
